requirements.txt
flask==3.1.1
langchain==0.3.27
chromadb==1.0.15
sentence-transformers==5.0.0
huggingface-hub==0.33.4
python-dotenv==1.1.1
langchain-huggingface==0.3.1
langchain-community==0.3.27
langchain-text-splitters==0.3.9
langchain-groq==0.3.6
langchain-chroma==0.2.5
langgraph==0.6.3
langchain-core==0.3.72
pymongo==4.14.0
libmagic==1.0